перенес шаги из тестов в LoginClient
