+++ src/test/java/tests/RegistrationTests.java
@@ -1,10 +1,9 @@
 package tests;
 
+import clients.LoginClient;
 import com.codeborne.selenide.Configuration;
 import com.codeborne.selenide.WebDriverRunner;
-import io.qameta.allure.Step;
 import io.qameta.allure.junit4.DisplayName;
-import io.restassured.http.ContentType;
 import io.restassured.response.Response;
 import models.Login;
 import models.LoginResponse;
@@ -20,9 +19,7 @@ import java.util.Objects;
 import java.util.Random;
 
 import static com.codeborne.selenide.Selenide.open;
-import static io.restassured.RestAssured.given;
 import static org.apache.http.HttpStatus.SC_ACCEPTED;
-import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertEquals;
 
 public class RegistrationTests {
@@ -55,13 +52,13 @@ public class RegistrationTests {
     @After
     public void clear() {
         if (Objects.equals(WebDriverRunner.getWebDriver().getCurrentUrl(), LOGIN_URL)) {
-            Login loginObject = createObjectLogin(email, password);
-            Response responseLoginCourier = sendPostRequestAuthLogin(loginObject);
-            LoginResponse loginResponse = deserialization(responseLoginCourier);
+            Login loginObject = LoginClient.createObjectLogin(email, password);
+            Response responseLoginCourier = LoginClient.sendPostRequestAuthLogin(loginObject);
+            LoginResponse loginResponse = LoginClient.deserialization(responseLoginCourier);
             String accessToken = loginResponse.getAccessToken();
 
-            Response responseDeleteCourier = sendDeleteRequestAuthUser(accessToken);
-            checkExpectedResult(responseDeleteCourier, SC_ACCEPTED, EXPECTED_RESULT_TRUE);
+            Response responseDeleteCourier = LoginClient.sendDeleteRequestAuthUser(accessToken);
+            LoginClient.checkExpectedResult(responseDeleteCourier, SC_ACCEPTED, EXPECTED_RESULT_TRUE);
         }
     }
 
@@ -86,32 +83,5 @@ public class RegistrationTests {
         registerPage.setPassword(shortPassword);
         registerPage.clickRegisterButton();
         registerPage.compareText(INVALID_PASSWORD);
-    }
-
-    @Step("Проверка соответствия ожидаемого результата")
-    public void checkExpectedResult(Response response, int statusCode, boolean expectedResult) {
-        response.then().assertThat().statusCode(statusCode).and().body("success", equalTo(expectedResult));
-    }
-
-    @Step("Создание объекта логин")
-    public Login createObjectLogin(String email, String password) {
-        return new Login(email, password);
-    }
-
-    @Step("Десериализация ответа на логин пользователя")
-    public LoginResponse deserialization(Response responseLoginUser) {
-        return responseLoginUser.as(LoginResponse.class);
-    }
-
-    @Step("Отправка POST запроса на /api/auth/login")
-    public static Response sendPostRequestAuthLogin(Login login) {
-        return given().contentType(ContentType.JSON).body(login)
-                .post("https://stellarburgers.nomoreparties.site/api/auth/login");
-    }
-
-    @Step("Отправка DELETE запроса на /api/auth/user")
-    public static Response sendDeleteRequestAuthUser(String accessToken) {
-        return given().header("Authorization", accessToken)
-                .delete("https://stellarburgers.nomoreparties.site/api/auth/user");
     }
 }
